Rename common package to decorators
