+++ server/bulletin/controllers/bullet.py
@@ -1,5 +1,5 @@
 from bulletin import app, db
-from bulletin.common import auth, validation
+from bulletin.decorators import auth, validation
 from bulletin.libs.bullet import propogate_bullet_id_update
 from bulletin.models.bullet import Bullet
 from bulletin.models.membership import RoleType
